requirements.txt
# Core Dependencies
python-telegram-bot>=20.7
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0
pydantic>=2.5.0
structlog>=23.2.0
tenacity>=8.2.3

# Database & Caching
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0  # PostgreSQL async driver
redis>=5.0.0
aiocache>=0.12.0

# Data Analysis & Visualization
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
numpy>=1.25.0

# Security & Authentication
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# Monitoring & Logging
sentry-sdk>=1.38.0
prometheus-client>=0.19.0

# Development Dependencies
# Code Quality
ruff>=0.1.6
black>=23.11.0
mypy>=1.7.0
pre-commit>=3.5.0
bandit>=1.7.5
safety>=2.3.0
vulture>=2.10  # Dead code detection
interrogate>=1.5.0  # Docstring coverage
pyupgrade>=3.15.0  # Syntax modernization
radon>=6.0.1  # Code complexity metrics

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
coverage>=7.3.0
codecov>=2.1.0
factory-boy>=3.3.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.25.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Type Stubs
types-requests>=2.31.0
types-redis>=4.6.0
types-pyjwt>=1.7.0

# Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0
pydantic-settings>=2.1.0
schedule>=1.2.0
